OpenAI/requirements.txt
openai==1.69.0
python-dotenv==1.1.0